# openapi.yaml
openapi: 3.0.3
info:
  title: CatWAF API
  version: 1.0.1
  description: |
    CatWAF's control-plane API — everything the dashboard itself calls.
    There's no separate "public API" here; this is the real thing the
    frontend uses, documented so you can script against it or build your
    own tooling on top.

    **Scope of this document:** CatWAF has ~167 routes across 19 files.
    This spec covers the core groups in real depth — auth and two-factor
    login, WAF engine settings, custom rules, IP/geo lists, the security
    score, honeypots, sensitive-file protection, operational modes
    (panic/maintenance/config signing), and the settings, protection and
    operations layers — rather than a shallow stub for every endpoint.
    Routes not yet documented here still exist and work; see
    `backend/routes/` for the full list. Expanding this file's coverage is
    tracked in `ROADMAP.md`.

    **The settings layer is self-describing.** Rather than enumerating ~290
    fields here, `GET /api/settings/schema` returns the live schema the
    backend validates against and the dashboard builds its controls from —
    field names, types, defaults, permitted values and help text. That is
    the authoritative reference; `docs/settings.md` is generated from it.

    **One thing worth knowing:** the dashboard stats/traffic endpoints
    (`/api/stats`, `/api/traffic/*`) return real data — sourced from
    Coraza's own audit log via `services/requestLog.js`, not a demo feed —
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Obtained from POST /api/auth/login. Unauthenticated requests are
        still accepted for GET routes (treated as the built-in `guest`/
        viewer role) — the token only matters for anything that mutates
        state, which requires the `admin` role.
...
paths:
...
  /api/settings/schema:
    get:
      tags: [Settings]
      summary: The live settings schema
      description: >
        Every group and field the backend validates against, with types,
        defaults, permitted values and help text. The dashboard builds its
        controls from this at runtime, so a setting added to the schema
        appears in the UI with no matching frontend change. Secret fields are
        marked write_only and their values are never included.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/settings:
    get:
      tags: [Settings]
      summary: Every group's current values, redacted
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/settings/{group}:
    parameters:
      - { name: group, in: path, required: true, schema: { type: string }, description: 'A group name from /api/settings/schema, e.g. access, challenge, tls' }
    get:
      tags: [Settings]
      summary: One group's schema and current values
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '404': { description: Unknown settings group }
    patch:
      tags: [Settings]
      summary: Change one or more fields in a group
      description: >
        Validated, written, rendered into the Caddyfile and validated by Caddy
        before anything is applied. If the result would not load, the previous
        configuration is kept and the response describes which phase failed.
        The response includes render_report.skipped — settings that are on but
        could not be rendered, usually because the installed Caddy lacks an
        optional module.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: Applied }
        '400': { description: Validation failed, or the rendered configuration was rejected and rolled back }
        '404': { description: Unknown settings group }
  /api/settings/{group}/preview:
    parameters:
      - { name: group, in: path, required: true, schema: { type: string } }
    post:
      tags: [Settings]
      summary: Diff the Caddyfile this change would produce, without applying it
      description: >
        Runs the same mutate-and-render pipeline as PATCH, stops before
        validation and reload, diffs the two rendered Caddyfiles, then restores
        state. Nothing is written. render_report.newly_skipped names anything
        the change would enable that could not actually take effect.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: OK }
        '400': { description: Validation failed }
  /api/settings/{group}/reset:
    parameters:
      - { name: group, in: path, required: true, schema: { type: string } }
    post:
      tags: [Settings]
      summary: Restore a group's defaults
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Reset }
  /api/config/render-report:
    get:
      tags: [Settings]
      summary: What was left out of the generated configuration, and why
      description: >
        Settings that are switched on but could not be rendered — a missing
        optional Caddy module or an unmet prerequisite. CatWAF skips such a
        directive rather than emitting one Caddy would reject, which would take
        the site down; this endpoint is how that decision is surfaced instead
        of being a silent no-op.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  # ─── Protection ────────────────────────────────────────────────────────
...